requirements.txt
# LangGraph and LangChain dependencies
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-groq>=0.0.1

# HTTP server for API
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0

# Environment and utilities
python-dotenv>=1.0.0
requests>=2.31.0

# Development dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1
